+++ VoxelZombieServer/Assets/IChunk.cs
@@ -2,7 +2,7 @@ using System;
 
 public interface IChunk
 {
-    UInt16 this[int x, int y, int z] { get; set; }
+    UInt64 this[int x, int y, int z] { get; set; }
 
     bool dirty { get; set; }
 }
